+++ examples/flow_filtering/main.c
@@ -58,7 +58,7 @@
 
 static volatile bool force_quit;
 
-static uint8_t port_id;
+static uint16_t port_id;
 static uint16_t nr_queues = 5;
 static uint8_t selected_queue = 1;
 struct rte_mempool *mbuf_pool;
